+++ app/components/dashboard-components/StatCardProps.tsx
